Add benchmarks for coord proxy generation
Benchmarks creation of a coords proxy from a Dataset and slice of a
Dataset.

Re #558
